+++ src/main/java/net/lzdq/winterbridge/WinterBridge.java
@@ -1,32 +1,28 @@
 package net.lzdq.winterbridge;
 
 import com.mojang.logging.LogUtils;
-import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.event.server.ServerStartingEvent;
-import net.minecraftforge.eventbus.api.IEventBus;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.ModLoadingContext;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import net.minecraftforge.fml.config.ModConfig.Type;
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.ModContainer;
+import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.config.ModConfig.Type;
+import net.neoforged.neoforge.common.NeoForge;
+import net.neoforged.neoforge.event.server.ServerStartingEvent;
 import org.slf4j.Logger;
 
 @Mod(WinterBridge.MODID)
 public class WinterBridge {
 	public static final String MODID = "winterbridge";
 	public static final Logger LOGGER = LogUtils.getLogger();
-	public WinterBridge(){
-		IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
-		//modEventBus.addListener(this::commonSetup);
-		MinecraftForge.EVENT_BUS.register(this);
-		//modEventBus.addListener(this::addCreative);
-		ModLoadingContext.get().registerConfig(Type.CLIENT, ModConfig.CONFIG, "winterbridge-client.toml");
+
+	// NeoForge injects the mod event bus and mod container into the constructor.
+	public WinterBridge(IEventBus modEventBus, ModContainer modContainer) {
+		NeoForge.EVENT_BUS.register(this);
+		modContainer.registerConfig(Type.CLIENT, ModConfig.CONFIG, "winterbridge-client.toml");
 	}
 
 	@SubscribeEvent
-	public void onServerStarting(ServerStartingEvent event)
-	{
-		// Do something when the server starts
+	public void onServerStarting(ServerStartingEvent event) {
 		LOGGER.info("HELLO from server starting");
 	}
 }